Allow using HSMs rather than forcing them to live on disk
These changes were taken from the phrack fork of sbsigntools and applied
to the repo that Linux distributions use.  The README.md in their repo
has the defails, but the short version is that these changes enable
dynamic engine loading so engines like cloudhsm can be used, and keeps the
engine in scope until after the PKCS7 operations.

References:
https://github.com/phrack/sbsigntools
https://git.kernel.org/pub/scm/linux/kernel/git/jejb/sbsigntools.git/